Optimize scheduler algorithm

+++ pkg/device/types.go
@@ -634,7 +634,7 @@ func (dev *Device) Score() float64 {
 	numRatio := util.SafeDiv(float64(dev.AllocatableNumber()), float64(dev.GetTotalNumber()))
 	memRatio := util.SafeDiv(float64(dev.AllocatableMemory()), float64(dev.GetTotalMemory()))
 	coreRatio := util.SafeDiv(float64(dev.AllocatableCores()), float64(dev.GetTotalCores()))
-	return (numRatio + memRatio + coreRatio) / 3.0 * util.HundredCore
+	return (numRatio + memRatio + coreRatio) / 3.0 * float64(util.HundredCore)
 }
 
 // GetPodDeviceClaim Retrieve device claim information for a pod,
